Add shared object smearing to marker table
Adds smearing to the marker table for shared objects. Objects are
smeared only if they appear as a mutable input to the transaction.
Additionally changes the way we notify the transaction manager so we
only notify if when we smear (i.e., we don't send synthetic notifications unless
the deleted shared object appears as a mutable input to the
transaction). This is important for proper sequencing of transactions.

Adds deleted objects that are used in transactions after the deletion
has occured in the `shared_objects` transaction effects field with an
object digest of OBJECT_DIGEST_DELETED.

Added a test to check interleave mutations and reads after an SO
deletion and make sure we track transaction dependencies correctly. Also
adds Mark's test.

More test will be written here shortly, but I believe these changes
should encompass all of the core changes needed.